+++ torch_points3d/models/base_model.py
@@ -29,6 +29,7 @@ class BaseModel(torch.nn.Module, TrackerInterface, DatasetInterface, CheckpointI
     """
 
     __REQUIRED_DATA__: List[str] = []
+    __REQUIRED_LABELS__: List[str] = []
 
     def __init__(self, opt):
         """Initialize the BaseModel class.
@@ -391,14 +392,22 @@ class BaseModel(torch.nn.Module, TrackerInterface, DatasetInterface, CheckpointI
     def cuda(self):
         return self.to(torch.device("cuda"))
 
-    def verify_data(self, data):
-        """ Goes through the __REQUIRED_DATA__ attribute of the model
+    def verify_data(self, data, forward_only=False):
+        """ Goes through the __REQUIRED_DATA__ and __REQUIRED_LABELS__ attribute of the model
         and verifies that the passed data object contains all required members.
         If something is missing it raises a KeyError exception.
         """
-        for attr in self.__REQUIRED_DATA__:
+        missing_keys = []
+        required_attributes = self.__REQUIRED_DATA__
+        if not forward_only:
+            required_attributes += self.__REQUIRED_LABELS__
+        for attr in required_attributes:
             if not hasattr(data, attr) or data[attr] is None:
-                raise KeyError("Missing attribute in your data object: %s. The model will fail to forward." % attr)
+                missing_keys.append(attr)
+        if len(missing_keys):
+            raise KeyError(
+                "Missing attributes in your data object: {}. The model will fail to forward.".format(missing_keys)
+            )
 
 
 class BaseInternalLossModule(torch.nn.Module):